feat(issue-1757): promo_points yuan → cents (1 点 = 1 分)
Systematic unit migration: users.promo_points stored yuan → cents.

- model: PromoPoints float64 decimal → Cents bigint (JSON cents)
- migration 20260824003 (up/down): existing yuan ×100 → cents
- writes: registration gift (99 元 → 9900), referral reg bonus,
  refund rebate A2, referral commission, gift deduction at order,
  payment callback deduct (raw gift_used cents direct), admin edit
- reads: wallet API / users/me / staff profile auto-cents via model
- user_rental gift_points_max: cents (total yuan × ratio × 100)
- frontend: no change needed (already cents-based /100 display;
  MembershipCenter shows 9900 点)
- tests: registration gift 9900, wallet promo_points/max_gift_amount
  cents; updated gift registration/session assertions
- docs: api.md §8.11 cents contract, cases.md L-05 unit note

DB migration execution on prerelease/production requires user
authorization (red line) — deferred.

Model: deepseek/deepseek-v4-flash

+++ backend/handlers/payment_calculate.go
@@ -29,9 +29,9 @@ type PaymentCalculateResponse struct {
 }
 
 type WalletInfo struct {
-	PromoPoints   float64 `json:"promo_points"`
-	MaxGiftRatio  float64 `json:"max_gift_ratio"`
-	MaxGiftAmount float64 `json:"max_gift_amount"`
+	PromoPoints   float64 `json:"promo_points"`    // cents (1 点 = 1 分, #1757)
+	MaxGiftRatio  float64 `json:"max_gift_ratio"`  // ratio, unchanged
+	MaxGiftAmount float64 `json:"max_gift_amount"` // cents
 }
 
 func CalculatePayment(c *gin.Context) {
@@ -101,7 +101,7 @@ func getWalletInfo(db *gorm.DB, userID, tenantID string, amount float64) (*Walle
 	}
 
 	return &WalletInfo{
-		PromoPoints:   user.PromoPoints,
+		PromoPoints:   float64(user.PromoPoints),
 		MaxGiftRatio:  maxGiftRatio,
 		MaxGiftAmount: math.Floor(amount * maxGiftRatio * 100 / 100),
 	}, nil
